Update module list handling
Functions like multiscan() and _subscan() now take a list of module
names instead of full paths.

Renamed MODULESLIST -> MODULE_LIST.

Stopped using globals as default parameters in certain functions so the
functions will use the most up-to-date version of the global instead of
the global's state when the function was defined.